[pdex] Finish unit test
