+++ blockchain/pdex/params.go
@@ -5,6 +5,7 @@ import (
 	"reflect"
 
 	"github.com/incognitochain/incognito-chain/common"
+	"github.com/incognitochain/incognito-chain/config"
 	"github.com/incognitochain/incognito-chain/dataaccessobject/statedb"
 )
 
@@ -26,6 +27,7 @@ func NewParams() *Params {
 		FeeRateBPS:               map[string]uint{},
 		PDEXRewardPoolPairsShare: map[string]uint{},
 		StakingPoolsShare:        map[string]uint{},
+		StakingRewardTokens:      []common.Hash{},
 	}
 }
 
@@ -107,4 +109,26 @@ func isValidPdexv3Params(
 
 func (params *Params) IsZeroValue() bool {
 	return reflect.DeepEqual(params, NewParams()) || params == nil
+}
+
+func (params *Params) readConfig() *Params {
+	res := &Params{
+		DefaultFeeRateBPS:               config.Param().PDexParams.Params.DefaultFeeRateBPS,
+		PRVDiscountPercent:              config.Param().PDexParams.Params.PRVDiscountPercent,
+		TradingProtocolFeePercent:       config.Param().PDexParams.Params.TradingProtocolFeePercent,
+		TradingStakingPoolRewardPercent: config.Param().PDexParams.Params.TradingStakingPoolRewardPercent,
+		StakingPoolsShare:               config.Param().PDexParams.Params.StakingPoolsShare,
+		MintNftRequireAmount:            config.Param().PDexParams.Params.MintNftRequireAmount,
+		MaxOrdersPerNft:                 config.Param().PDexParams.Params.MaxOrdersPerNft,
+	}
+	if res.FeeRateBPS == nil {
+		res.FeeRateBPS = make(map[string]uint)
+	}
+	if res.StakingPoolsShare == nil {
+		res.StakingPoolsShare = make(map[string]uint)
+	}
+	if res.PDEXRewardPoolPairsShare == nil {
+		res.PDEXRewardPoolPairsShare = make(map[string]uint)
+	}
+	return res
 }
